Commit introduces verbose path calling
This commit cleans up a bit and adds full path names to the function
calls in the lib.rs file. We do this to be explicit about the calls
we are making to avoid overlap in the user space when lovely people
use this crate in their code.

Changes to be committed:
  modified:   src/lib.rs